Se refactoriza los adapters por carpetas de agregados
