+++ back-springboot/infrastructure/driven-adapters/mongo-repository/src/main/java/co/com/sofka/mongo/count/CountMongoDBRepository.java
@@ -1,4 +1,4 @@
-package co.com.sofka.mongo;
+package co.com.sofka.mongo.count;
 
 import co.com.sofka.mongo.entity.CountEntity;
 import org.springframework.data.mongodb.repository.MongoRepository;
