+++ back-springboot/infrastructure/driven-adapters/mongo-repository/src/main/java/co/com/sofka/mongo/count/CountMongoRepositoryAdapter.java
@@ -1,4 +1,4 @@
-package co.com.sofka.mongo;
+package co.com.sofka.mongo.count;
 
 import co.com.sofka.model.count.Count;
 import co.com.sofka.model.count.gateways.CountRepository;
